requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
av==13.1.0
bitsandbytes==0.44.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.1.0
decord==0.6.0
deepspeed==0.15.4
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.7/flash_attn-2.5.7+cu122torch2.1cxx11abiFALSE-cp39-cp39-linux_x86_64.whl#sha256=09e0e7cb90d6c043b5a86fee184333b43b86311e21ec1f67380d7b4141d642d3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.31.5
gradio_client==0.16.4
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.11.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.14.2
opencv-python==4.10.0.84
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pathos==0.3.2
peft==0.13.2
Pillow==9.3.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
prettytable==3.10.0
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
ruff==0.7.1
safetensors==0.4.5
scikit-learn==1.4.1.post1
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk @ file:///home/panmx/sentry_sdk-2.18.0-py2.py3-none-any.whl#sha256=ee70e27d1bbe4cd52a38e1bd28a5fadb9b17bc29d91b5f2b97ae29c0a7610442
setproctitle==1.3.4
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.21.0
tomlkit==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.5
transformers==4.47.0
triton==2.1.0
typeguard==4.4.1
typer==0.12.5
typing_extensions==4.12.2
tyro==0.9.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wandb==0.18.7
wcwidth==0.2.13
websockets==11.0.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.20.2
